+++ src/model/event-type.entity.ts
@@ -16,7 +16,7 @@ export class EventType extends BaseEntity {
   @Column({ type: 'enum', enum: EventTypeTrigger, nullable: true })
   trigger: EventTypeTrigger;
 
-  @Column({ type: 'float', nullable: true })
+  @Column({ type: 'numeric', nullable: true })
   penaltyValue: number;
 
   @Column({ type: 'enum', enum: PenaltyUnit, nullable: true })
